Created unit_selection class
